widen control values from Float to Double across the FFI
This is an ABI break: clients that were holding a stale shared
library will silently corrupt control args. The Haskell side now
carries Double through Connection (Param), InputConn (Literal),
RuntimeInput (RConst) and the rnControls/irControls vectors;
c_rt_graph_set_control marshals CDouble. The C++ side stores
node.controls as std::vector<double> and downcasts to float at
the multiply sites in process_gain / process_add.

Audio sample buffers stay float — the widening is purely on the
control path, where q's frequency/duration/config constructors
take double anyway and the previous code was casting back at
every kernel call.

Test generators in test/Spec.hs are updated in lockstep so the
QC properties keep checking the right shape.

+++ test/Spec.hs
@@ -320,11 +320,11 @@ propValidates g = case validateAndSort g of
 -- the very start, before any source has been added.
 
 data Op
-  = OSinOsc Float Float
-  | OSawOsc Float Float
+  = OSinOsc Double Double
+  | OSawOsc Double Double
   | ONoise
-  | OGain   Int Float        -- source-index, gain amount
-  | OLPF    Int Float Float  -- source-index, cutoff, q
+  | OGain   Int Double       -- source-index, gain amount
+  | OLPF    Int Double Double  -- source-index, cutoff, q
   | OOut    Int Int          -- channel, source-index
   deriving (Eq, Show)
